+++ Connect/Services/ChatHub.cs
@@ -1,32 +1,14 @@
-﻿using Connect.Models;
-using Microsoft.AspNet.SignalR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System;
 using System.Web;
-
-namespace Connect.Services
+using Microsoft.AspNet.SignalR;
+namespace SignalRChat
 {
     public class ChatHub : Hub
     {
-        static List<User> ConnectedUsers = new List<User>();
-        static List<Message> CurrentMessage = new List<Message>();
-
-        public void Connect(string userName)
+        public void Send(string name, string message)
         {
-            var id = Context.ConnectionId;
-
-            if (ConnectedUsers.Count(x => x.ConnectionID == id) == 0)
-            {
-                string logintime = DateTime.Now.ToString();
-                ConnectedUsers.Add(new User { ConnectionID = id, Username = userName});
-
-                // send to caller  
-                Clients.Caller.onConnected(id, userName, ConnectedUsers, CurrentMessage);
-
-                // send to all except caller client  
-                Clients.AllExcept(id).onNewUserConnected(id, userName);
-            }
+            // Call the broadcastMessage method to update clients.
+            Clients.All.broadcastMessage(name, message);
         }
     }
 }
