+++ Connect/Site.Master
@@ -15,9 +15,20 @@
     <link href="~/favicon.ico" rel="shortcut icon" type="image/x-icon" />
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
 
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+ <%--   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>--%>
+
+     <script src="../Scripts/jquery-3.2.1.min.js" ></script>
+    <!--Reference the SignalR library. -->
+    <script src="../Scripts/jquery.signalR-2.2.2.min.js"></script>
+
+    <%--<script src="../Scripts/jquery.signalR-2.2.2.min.js"></script>--%>
 
     <link rel="stylesheet" type="text/css" href="./Css/main.css">
+
+    <!--Reference the autogenerated SignalR hub script. -->
+    <%--<script src="~/signalr/hubs"></script>--%>
+    <script src='<%: ResolveClientUrl("~/signalr/hubs") %>'></script>
+
 </head>
 <body>
     <div runat="server">
