+++ Connect/Startup.cs
@@ -1,17 +1,13 @@
 ﻿using Microsoft.Owin;
 using Owin;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-[assembly: OwinStartup(typeof(Connect.Startup))]
-namespace Connect
+[assembly: OwinStartup(typeof(SignalRChat.Startup))]
+namespace SignalRChat
 {
     public class Startup
     {
         public void Configuration(IAppBuilder app)
         {
+            // Any connection or hub wire up and configuration should go here
             app.MapSignalR();
         }
     }
