+++ Connect/packages.config
@@ -2,9 +2,9 @@
 <packages>
   <package id="Antlr" version="3.5.0.2" targetFramework="net472" />
   <package id="AspNet.ScriptManager.bootstrap" version="3.4.1" targetFramework="net472" />
-  <package id="AspNet.ScriptManager.jQuery" version="3.4.1" targetFramework="net472" />
+  <package id="AspNet.ScriptManager.jQuery" version="3.2.1" targetFramework="net472" />
   <package id="bootstrap" version="3.4.1" targetFramework="net472" />
-  <package id="jQuery" version="3.4.1" targetFramework="net472" />
+  <package id="jQuery" version="3.2.1" targetFramework="net472" />
   <package id="Microsoft.AspNet.FriendlyUrls" version="1.0.2" targetFramework="net472" />
   <package id="Microsoft.AspNet.FriendlyUrls.Core" version="1.0.2" targetFramework="net472" />
   <package id="Microsoft.AspNet.ScriptManager.MSAjax" version="5.0.0" targetFramework="net472" />
